requirements.txt
# Web server
flask

# Database
mariadb

# Utilities
python-dotenv
requests
beautifulsoup4
apscheduler

# HTTP client with browser TLS fingerprinting (primary fetcher)
curl-cffi

# Testing
pytest
